Host list markup and less changes (WIP)
Modified the host list markup to reflect host information.
Comments are currently missing as well as behaviour and final styling.
The styling part is not really part of the ticket, as we decided
not to style or add additional markup, but as the work
has been done in sprint 1.0-4 it makes no sense to
throw it away at this time

refs #4179