list/hosts: show unhandled services righthand
This way we automagically fix the problem with the wrong row links.
I guess we should re-introduce the "row-action" class allowing "precedence"
rules, so also a link not being the first one in a row could act as "row-link".

Changed formatting / text output a little bit, IMO even more cure is needed.

fixes #6645