Properly build the unhandled action links when showing multiple objects
The links designated to acknowledge unhandled problems or to schedule
a downtime for them were generated based on the full list of objects and
limited by non-supported filter parameters. As we are already aware of the
exact "unhandled" objects this list is now used to generate proper links.
(Btw applying filters to URLs is a mess...)

fixes #8017