Mute all icons that are part of an anchor and add simple labels to such anchors
The labels I've added are probably just of temporary nature as it's not
defined yet how to handle link labels properly for screen readers. What
is sure is that all icons that are part of a link are not important for the
meaning of it as this is expressed by the link itself.

refs #8360
refs #8358