Use timeSince() for displaying status changes of services instead of prefixedTimeSince()
Note that only the services view excluded the prefix from time diffs when being displayed in the dashboard. I'll readd that for both hosts and services later on.
Further, note that only the hosts view excludes the date of the status change when in state pending. I'll add that for the services view too.

refs #6778